Added test for JIT compilation
