+++ MarkMpn.Sql4Cds.Engine.Tests/Sql2FetchXmlTests.cs
@@ -2020,7 +2020,6 @@ namespace MarkMpn.Sql4Cds.Engine.Tests
         {
             var query = "SELECT displayname FROM metadata.globaloptionset WHERE name = 'test'";
 
-            _context.AddFakeMessageExecutor<RetrieveAllOptionSetsRequest>(new RetrieveAllOptionSetsHandler());
             var planBuilder = new ExecutionPlanBuilder(_localDataSource.Values, this);
             var queries = planBuilder.Build(query, null, out _);
 
@@ -2045,7 +2044,6 @@ namespace MarkMpn.Sql4Cds.Engine.Tests
         {
             var query = "SELECT logicalname FROM metadata.entity ORDER BY 1";
 
-            _context.AddFakeMessageExecutor<RetrieveMetadataChangesRequest>(new RetrieveMetadataChangesHandler(_localDataSource["local"].Metadata));
             var planBuilder = new ExecutionPlanBuilder(_localDataSource.Values, this);
             var queries = planBuilder.Build(query, null, out _);
 
@@ -2070,7 +2068,6 @@ namespace MarkMpn.Sql4Cds.Engine.Tests
         {
             var query = "SELECT e.logicalname, a.logicalname FROM metadata.entity e INNER JOIN metadata.attribute a ON e.logicalname = a.entitylogicalname WHERE e.logicalname = 'new_customentity' ORDER BY 1, 2";
 
-            _context.AddFakeMessageExecutor<RetrieveMetadataChangesRequest>(new RetrieveMetadataChangesHandler(_localDataSource["local"].Metadata));
             var planBuilder = new ExecutionPlanBuilder(_localDataSource.Values, this);
             var queries = planBuilder.Build(query, null, out _);
 
@@ -2593,212 +2590,6 @@ namespace MarkMpn.Sql4Cds.Engine.Tests
 
         void IQueryExecutionOptions.ConfirmDelete(ConfirmDmlStatementEventArgs e)
         {
-        }
-
-        private class RetrieveMetadataChangesHandler : IFakeMessageExecutor
-        {
-            private readonly IAttributeMetadataCache _metadata;
-
-            public RetrieveMetadataChangesHandler(IAttributeMetadataCache metadata)
-            {
-                _metadata = metadata;
-            }
-
-            public bool CanExecute(OrganizationRequest request)
-            {
-                return request is RetrieveMetadataChangesRequest;
-            }
-
-            public OrganizationResponse Execute(OrganizationRequest request, XrmFakedContext ctx)
-            {
-                var metadata = new EntityMetadataCollection
-                {
-                    _metadata["account"],
-                    _metadata["contact"],
-                    _metadata["new_customentity"]
-                };
-
-                foreach (var entity in metadata)
-                {
-                    if (entity.MetadataId == null)
-                        entity.MetadataId = Guid.NewGuid();
-
-                    foreach (var attribute in entity.Attributes)
-                    {
-                        if (attribute.MetadataId == null)
-                            attribute.MetadataId = Guid.NewGuid();
-                    }
-                }
-
-                var req = (RetrieveMetadataChangesRequest)request;
-                var metadataParam = Expression.Parameter(typeof(EntityMetadata));
-                var filter = ToExpression(req.Query.Criteria, metadataParam);
-                var filterFunc = (Func<EntityMetadata,bool>) Expression.Lambda(filter, metadataParam).Compile();
-
-                var result = new EntityMetadataCollection();
-
-                foreach (var match in metadata.Where(e => filterFunc(e)))
-                    result.Add(match);
-
-                var response = new RetrieveMetadataChangesResponse
-                {
-                    Results = new ParameterCollection
-                    {
-                        ["EntityMetadata"] = result
-                    }
-                };
-
-                return response;
-            }
-
-            private Expression ToExpression(MetadataFilterExpression filter, ParameterExpression param)
-            {
-                if (filter == null)
-                    return Expression.Constant(true);
-
-                Expression expr = null;
-
-                foreach (var condition in filter.Conditions)
-                {
-                    var conditionExpr = ToExpression(condition, param);
-
-                    if (expr == null)
-                        expr = conditionExpr;
-                    else if (filter.FilterOperator == LogicalOperator.And)
-                        expr = Expression.AndAlso(expr, conditionExpr);
-                    else
-                        expr = Expression.OrElse(expr, conditionExpr);
-                }
-
-                foreach (var subFilter in filter.Filters)
-                {
-                    var filterExpr = ToExpression(subFilter, param);
-
-                    if (expr == null)
-                        expr = filterExpr;
-                    else if (filter.FilterOperator == LogicalOperator.And)
-                        expr = Expression.AndAlso(expr, filterExpr);
-                    else
-                        expr = Expression.OrElse(expr, filterExpr);
-                }
-
-                return expr ?? Expression.Constant(true);
-            }
-
-            private Expression ToExpression(MetadataConditionExpression condition, ParameterExpression param)
-            {
-                var value = Expression.PropertyOrField(param, condition.PropertyName);
-                var targetValue = Expression.Constant(condition.Value);
-
-                switch (condition.ConditionOperator)
-                {
-                    case MetadataConditionOperator.Equals:
-                        return Expression.Equal(value, targetValue);
-
-                    case MetadataConditionOperator.NotEquals:
-                        return Expression.NotEqual(value, targetValue);
-
-                    case MetadataConditionOperator.LessThan:
-                        return Expression.LessThan(value, targetValue);
-
-                    case MetadataConditionOperator.GreaterThan:
-                        return Expression.GreaterThan(value, targetValue);
-
-                    default:
-                        throw new NotImplementedException();
-                }
-            }
-
-            public Type GetResponsibleRequestType()
-            {
-                return typeof(RetrieveMetadataChangesRequest);
-            }
-        }
-
-        private class RetrieveAllOptionSetsHandler : IFakeMessageExecutor
-        {
-            public bool CanExecute(OrganizationRequest request)
-            {
-                return request is RetrieveAllOptionSetsRequest;
-            }
-
-            public OrganizationResponse Execute(OrganizationRequest request, XrmFakedContext ctx)
-            {
-                var labels = new[]
-                {
-                    new LocalizedLabel("TestGlobalOptionSet", 1033) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("TranslatedDisplayName-Test", 9999) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("FooGlobalOptionSet", 1033) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("TranslatedDisplayName-Foo", 9999) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("BarGlobalOptionSet", 1033) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("TranslatedDisplayName-Bar", 9999) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("Value1", 1033) { MetadataId = Guid.NewGuid() },
-                    new LocalizedLabel("Value2", 1033) { MetadataId = Guid.NewGuid() }
-                };
-
-                return new RetrieveAllOptionSetsResponse
-                {
-                    Results = new ParameterCollection
-                    {
-                        ["OptionSetMetadata"] = new OptionSetMetadataBase[]
-                        {
-                            new OptionSetMetadata(new OptionMetadataCollection(new[]
-                            {
-                                new OptionMetadata(1) { MetadataId = Guid.NewGuid(), Label = new Label { UserLocalizedLabel = labels[6] } },
-                                new OptionMetadata(2) { MetadataId = Guid.NewGuid(), Label = new Label { UserLocalizedLabel = labels[7] } }
-                            }))
-                            {
-                                MetadataId = Guid.NewGuid(),
-                                Name = "test",
-                                DisplayName = new Label(
-                                    labels[0],
-                                    new[]
-                                    {
-                                        labels[0],
-                                        labels[1]
-                                    })
-                            },
-                            new OptionSetMetadata(new OptionMetadataCollection(new[]
-                            {
-                                new OptionMetadata(1) { MetadataId = Guid.NewGuid(), Label = new Label { UserLocalizedLabel = labels[6] } },
-                                new OptionMetadata(2) { MetadataId = Guid.NewGuid(), Label = new Label { UserLocalizedLabel = labels[7] } }
-                            }))
-                            {
-                                MetadataId = Guid.NewGuid(),
-                                Name = "foo",
-                                DisplayName = new Label(
-                                    labels[2],
-                                    new[]
-                                    {
-                                        labels[2],
-                                        labels[3]
-                                    })
-                            },
-                            new OptionSetMetadata(new OptionMetadataCollection(new[]
-                            {
-                                new OptionMetadata(1) { MetadataId = Guid.NewGuid(), Label = new Label { UserLocalizedLabel = labels[6] } },
-                                new OptionMetadata(2) { MetadataId = Guid.NewGuid(), Label = new Label { UserLocalizedLabel = labels[7] } }
-                            }))
-                            {
-                                MetadataId = Guid.NewGuid(),
-                                Name = "bar",
-                                DisplayName = new Label(
-                                    labels[4],
-                                    new[]
-                                    {
-                                        labels[4],
-                                        labels[5]
-                                    })
-                            }
-                        }
-                    }
-                };
-            }
-
-            public Type GetResponsibleRequestType()
-            {
-                return typeof(RetrieveAllOptionSetsRequest);
-            }
         }
     }
 }
